requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
black==24.3.0
boto3==1.34.48
botocore==1.34.162
CacheControl==0.14.4
cbor2==5.9.0
celery==5.4.0
certifi==2026.2.25
cffi==2.0.0
channels==4.1.0
channels-redis==4.1.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.6
daphne==4.1.0
Django==6.0.3
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-crispy-forms==2.1
django-environ==0.13.0
django-extensions==3.2.3
django-filter==25.2
django-ratelimit==4.1.0
django-storages==1.14.2
django-timezone-field==7.2.1
djangorestframework==3.14.0
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
firebase-admin==6.4.0
flake8==7.0.0
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.1
grpcio==1.78.0
grpcio-status==1.78.0
httplib2==0.31.2
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
isort==5.13.2
jmespath==1.1.0
kombu==5.6.2
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==3.0
pyflakes==3.2.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
pytest==7.4.4
pytest-django==4.7.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2026.1.post1
redis==7.4.0
requests==2.33.0
s3transfer==0.10.4
service-identity==24.2.0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
wheel==0.46.3
whitenoise==6.6.0
zope.interface==8.2
